Invalid Address Fix
- Fixed a bug where a valid address was showing up as invalid. This was caused because I was checking for and int instead of a long

+++ LiskExcel/Form1.cs
@@ -36,8 +36,8 @@ namespace LiskExcel
             }
             else
             {
-                int n;
-                bool isNumeric = int.TryParse(address.Substring(0, address.Length - 1), out n);
+                long n;
+                bool isNumeric = long.TryParse(address.Substring(0, address.Length - 1), out n);
                 return isNumeric;
             }
         }
